[Refactor] Move and restructure Landing components
