+++ src/pages/Landing/Landing.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Header } from "@/components/sections/Header";
-import { Hero } from "@/components/layouts/Landing/Hero";
+import { Hero } from "@/components/layouts/Landing/Hero/Hero";
 import { Frame } from "@/components/layouts/Landing/Frame";
-import { TechShowcase } from "@/components/layouts/Landing/Animated/TechShowcase";
-import { GameEngine } from "@/components/layouts/Landing/Animated/GameEngine";
+import { TechShowcase } from "@/components/layouts/Landing/TechShowcase";
+import { GameEngine } from "@/components/layouts/Landing/GameEngine";
 import { Footer } from "@/components/sections/Footer";
 
 const Landing = () => {
